generate one-time health error on loss of AC power
Obviously, this won't work if both power supplies go bad at once.

It will *not* generate an error every time there's no reading on one of
the AC Voltage sensors - they don't work on older x950 systems, and it
would be annoying to insist that customers attach both power supplies
when they work just fine with only one (albeit with less fault-tolerance).

Instead, it just generates the health error (once) when a power supply
that did have power stops working.

In addition to the voltage warnings, there are likely to be system event
log (SEL) messages that will be displayed - we don't generate the health
error based on those, as they can be caused by normal circumstances, e.g.
shutting down the system, unplugging it and moving it to a new rack
before plugging it back in and powering up again.